+++ js/util.js
@@ -1,3 +1,35 @@
+// Модуль содержит утилитарные функции
+
+// Функция генерирует случайное число из диапазона от min до max включительно
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// Функция создает генератор случайных чисел из диапазона от min до max включительно так, что число не повторяется, пока не будут перебраны все числа из диапазона
+const createRandomIntegerFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return () => {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+export {getRandomInteger, createRandomIntegerFromRangeGenerator};
+
+/*
+--- разобрать потом ---
+
 // Функция для проверки длины строки
 const checkStringLength = function (string, stringMaxLength) {
   const normalizeString = string.toString();
@@ -40,3 +72,5 @@ const getNumbers = function (string) {
 checkStringLength('проверяемая строка', 20);
 isPalindrom('Лёша на полке клопа нашёл ');
 getNumbers('ECMAScript 2022');
+
+*/
